backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# AI/LLM Integration
httpx>=0.25.2,<0.26.0
ollama==0.1.6

# Data & Export
pydantic==2.5.3
pydantic-settings==2.1.0
icalendar==5.0.11
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet==3.0.3

# Utilities
python-dotenv==1.0.0
